refactor(scripts): move validate-docs-pages out of .ai/scripts
The MDX docs-page check enforces a general project authoring standard,
not an AI-tooling-specific invariant. Move it next to the other
repo-wide command scripts (`scripts/publish.js`,
`scripts/generate-versions.js`) so the layout reflects its scope.

- `.ai/scripts/validate-docs-pages.js` → `scripts/validate-docs-pages.js`
- `repoRoot` recomputed: the file is now one level below the repo root
  instead of two
- `.ai/scripts/validate.js` orchestrator imports from
  `../../scripts/validate-docs-pages.js` so `yarn lint:ai` still bundles
  the check
- `package.json` `lint:docs-pages` script path updated
- `.husky/pre-commit` path updated
- Header comment in the script updated to reference the `yarn`
  commands rather than the old script path

The AI rules (`.ai/rules/stories-documentation.md`) and `.ai/README.md`
reference the `yarn lint:docs-pages` command, not the script path, so
they continue to work unchanged.
